Move service functions to a separate module
